test: split entity manager lifecycle tests
